allow custom meta fields and templates per page
any fields can be added into the page meta section for use in themes
e.g.
/*
BodyClass: amazeballs
*/
could be used in a template as:
<body{% if meta.bodyclass %} class="{{ meta.bodyclass }}"{% endif %}>

Specifying a "Template" field will attempt to load a a template html
file with that name from the current theme directory e.g.
/*
Template: spesh
*/
attempts to use "spesh.html" but falls back to "index.html" if the file
does not exist.