[FEATURE] Multi Template Support
this patch makes it possible to use different templates by define the template in the meta section of the *.md files. This is the first patch for multi column support.

To use an other template like index.html you can now define Template: content-sidebar which renders the template "content-sidebar.html".