Updated customization links everywhere.
I've searched the project and replaced all instances of `{{ site.github.url }}/customization` with the relevent `/plugins` or `/themes`, changing as little as possible.  I consider
most of the site content to be on "life-support" until it gets rewritten, but this will ensure that there are no dead links in the meantime.