Portfolio: Refactor JavaScript
With this mess of a code I wasn't able to track the bug down, so many useless and conflicting calls, that's impossible to understand. So I've decided to do a complete code refactoring. The result is a much smoother and easier-to-understand codebase that should be maintainable. Still far away from being perfect (both Isotope and jQuery fuck up the HTML code, thus it's basically impossible to get back to an "clean" state)... Not getting back to a "clean" state makes it quite impossible to make events abortable, so you'll have to wait until the closing animation is finished before the opening animation starts ("you" in terms of "your browser").

TODO: Test this with IE8+ (IE9+ is required) and mobile browsers...